Refactoring, TOdo: Redesign Play.sendtxn
